adding interface to listener

+++ Job_Manager_API.py
@@ -1,8 +1,8 @@
 import uuid
 from Job_Manager_Thread_Safe import Job_Manager_Thread_Safe
-from Bio import SeqIO
 import os
 import shutil
+from InputValidator import InputValidator
 
 
 class Job_Manager_API:
@@ -10,23 +10,7 @@ class Job_Manager_API:
         self.__input_file_name = input_file_name
         self.__upload_root_path = upload_root_path
         self.j_manager_thread_safe = Job_Manager_Thread_Safe(max_number_of_process, upload_root_path, input_file_name)
-
-    def __is_fasta(self, filename):
-        with open(filename, "r") as handle:
-            fasta = SeqIO.parse(handle, "fasta")
-            try:
-                return any(fasta)
-            except Exception as e:
-                return False
-            
-    def __is_fastq(self, filename):
-        with open(filename, "r") as handle:
-            fastq = SeqIO.parse(handle, "fastq")
-            
-            try:
-                return any(fastq)
-            except Exception as e:
-                return False
+        self.input_validator = InputValidator()
     
     def __delete_folder(self, process_id):
         folder2remove = os.path.join(self.__upload_root_path, process_id)
@@ -34,10 +18,11 @@ class Job_Manager_API:
     
     def __validate_input_file(self, process_id):
         parent_folder = os.path.join(self.__upload_root_path, process_id)
+        if not os.path.isdir(parent_folder):
+            print('Job_Manager_API', '__validate_input_file', f'no folder of {process_id}')
+            return False
         file2check = os.path.join(parent_folder, self.__input_file_name)
-        if self.__is_fasta(file2check):
-            return True
-        elif self.__is_fastq(file2check):
+        if self.input_validator.validate_input_file(file2check):
             return True
         self.__delete_folder(process_id)
         return False
@@ -57,9 +42,10 @@ class Job_Manager_API:
     def get_waiting_process(self):
         return self.j_manager_thread_safe.get_waiting_process()
         
-    def process_finished(self, process_id: str):
-        self.j_manager_thread_safe.change_process_state_finished(process_id)
-        process2add = self.j_manager_thread_safe.pop_from_waiting_queue()
-        self.j_manager_thread_safe.add_process(process_id)
+    def get_job_state(self, process_id):
+        state = self.j_manager_thread_safe.get_job_state(process_id)
+        if state:
+            return state
+        return 'job unavialiable'
